defult address update and user role update

+++ admin_panel/category.py
@@ -7,11 +7,12 @@ import re
 
 app = connect_db.app
 
-
+text = "admin"
 cors = CORS(app)
 app.config["UPLOAD_FOLDER"] = "static/img/"
 
 time = datetime.today().strftime("%Y-%m-%d")
+
 
 def admin_log(user_id, action, action_date, ip_address):
     conn = get_db_connection()
@@ -105,39 +106,45 @@ def get_all_category_filter(name, search, limit):
 
 # Create a new product
 def create_category(name, description, parent_category_id, created_at, picture):
-    conn = get_db_connection()
-    cur = conn.cursor()
-    cur.execute(
-        "INSERT INTO Categories (category_name,description, parent_category_id,created_at,picture) VALUES (?, ?, ?,?,?)",
-        (name, description, parent_category_id, created_at, picture),
-    )
-    conn.commit()
-    category_id = cur.lastrowid
-    conn.close()
-    return category_id
+    if text == "Admin" or text == "Super Admin":
+
+        conn = get_db_connection()
+        cur = conn.cursor()
+        cur.execute(
+            "INSERT INTO Categories (category_name,description, parent_category_id,created_at,picture) VALUES (?, ?, ?,?,?)",
+            (name, description, parent_category_id, created_at, picture),
+        )
+        conn.commit()
+        category_id = cur.lastrowid
+        conn.close()
+        return category_id
 
 
 # Update a product
 def update_category(name, description, parent_category_id, category_id):
-    conn = get_db_connection()
-    cur = conn.cursor()
-    cur.execute(
-        "UPDATE Categories SET category_name = ?,description = ?, parent_category_id = ? WHERE category_id = ?",
-        (name, description, parent_category_id, category_id),
-    )
-    conn.commit()
-    conn.close()
-    return get_category(category_id)
+    if text == "Admin" or text == "Super Admin":
+    
+        conn = get_db_connection()
+        cur = conn.cursor()
+        cur.execute(
+            "UPDATE Categories SET category_name = ?,description = ?, parent_category_id = ? WHERE category_id = ?",
+            (name, description, parent_category_id, category_id),
+        )
+        conn.commit()
+        conn.close()
+        return get_category(category_id)
 
 
 # Delete a product
 def delete_category(category_id):
-    conn = get_db_connection()
-    cur = conn.cursor()
-    cur.execute("DELETE FROM Categories WHERE category_id = ?", (category_id,))
-    cur.execute("DELETE FROM Products WHERE categories_id = ?", (category_id,))
-    conn.commit()
-    conn.close()
+    if text == "Admin" or text == "Super Admin":
+    
+        conn = get_db_connection()
+        cur = conn.cursor()
+        cur.execute("DELETE FROM Categories WHERE category_id = ?", (category_id,))
+        cur.execute("DELETE FROM Products WHERE categories_id = ?", (category_id,))
+        conn.commit()
+        conn.close()
 
 
 # CRUD routes
@@ -170,21 +177,22 @@ def list_category():
     return response
 
 
+
 @app.route("/category", methods=["POST"])
 def add_category():
-    name = request.json["name"]
-    description = request.json["description"]
-    parent_category_id = request.json["parent_category_id"]
+        name = request.json["name"]
+        description = request.json["description"]
+        parent_category_id = request.json["parent_category_id"]
 
-    picture = request.json["picture"]
-    user_ip = request.remote_addr
+        picture = request.json["picture"]
+        user_ip = request.remote_addr
 
-    action = f"Add Category: {name}"
-    admin_log(3, action, time, user_ip)
-    category_id = create_category(
-        name, description, parent_category_id, time, "picture.read()"
-    )
-    return jsonify(get_category(category_id)), 201
+        action = f"Add Category: {name}"
+        admin_log(3, action, time, user_ip)
+        category_id = create_category(
+            name, description, parent_category_id, time, "picture.read()"
+        )
+        return jsonify(get_category(category_id)), 201
 
 
 @app.route("/category/<int:category_id>", methods=["GET"])
@@ -193,6 +201,7 @@ def get_category_by_id(category_id):
     if category is None:
         return "", 404
     return jsonify(category), 200
+
 
 
 @app.route("/category/<int:category_id>", methods=["PUT"])
